Remove Console.Clear() and add comment about CreatedAt issue
Removed the Console.Clear() statement from Program.cs. Added a comment to highlight an issue with the user.CreatedAt property displaying incorrect information due to the addition of a datetime entity. This needs to be fixed in the future.

Remove Console.Clear() from Program.cs

The Console.Clear() statement was removed from the Program.cs file. This change modifies the behavior of the console application by no longer clearing the console screen before displaying the list of users. The rest of the code within the foreach loop remains unchanged, continuing to print user details to the console.